Update README.md in response to "Sample Image Issue #49
Hi,
This commit is in response to "Sample Image Issue #49" posted in the "weecology/DeepForest" repository's "Issues section". 

This image will let the user get a better practical real-time use-case of this python package. It will do so as this image has a combination of both trees (that too of different variations in height and crown size) and non-tree objects (to be specific vehicles, buildings and concrete flooring ). However, still from the image attached, it is very well clear that the python package has very accurately recognized only trees and marked them with bounding boxes. 

Thus, in short, this committed image will encourage more users to trust, get more comfortable using this python package in their projects and contributing in return towards the development of the DeepForest package.