Create deep copy for selecting annotations
preprocess::select_annotations is unclear about whether it is trying to
change `annotations` in place or create a new data frame. The return
docs indicate a data frame and given the kind of work that's happening it
appears that a new data frame is what we want.
This makes this explicit by making a deep copy after the subsetting.

This removes over 350 warnings from the tests.