Create deep copy when selecting annotations in preprocess (#436)
* Fix SettingWithCopy issues in tests

The way we were chaining subsets in Pandas can cause issues and was
therefore triggering warnings. Using .iloc once for all subsetting fixes
this issue:
https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#returning-a-view-versus-a-copy

* Create deep copy for selecting annotations

preprocess::select_annotations is unclear about whether it is trying to
change `annotations` in place or create a new data frame. The return
docs indicate a data frame and given the kind of work that's happening it
appears that a new data frame is what we want.
This makes this explicit by making a deep copy after the subsetting.

This removes over 350 warnings from the tests.