Automatically remove waiting label after new comment
We have added a new label - "Awaiting author contribution" - to indicate when we have responded to an issue/PR and are waiting for poster to respond.
This GH Action automatically removes the label when a new comment is added to the issue/PR.
This ensures that we safely ignore issues with that label and likely automatically close those issues after some period of time.