Ignore PyTorch PossibleUserWarnings during testing
There are a number of warnings in PyTorch that represent advice for
improving model, data loader, and training decisions. These warnings
are often triggered during testing. PyTorch supports disabling this
class of warnings, PossibleUserWarnings.

See:
https://lightning.ai/docs/pytorch/2.4.0/advanced/warnings.html
https://github.com/Lightning-AI/pytorch-lightning/issues/10644

This change follows the recommended approach to hide these warnings
during testing to avoid cluttering the logs. We were filtering several
of them manually, but those filters had stopped working due to changes
in the wording of the warning.

As of fa8f1df37327d0972bc41fa2c0db0503d13a2a8c these warnings are
suppressed for users, but pytest appears to override warnings filtered
by the package itself so we still need this.