Distribute config file from primary location
We had two copies of deepforest_config.yml. The first was the main copy
in the root directory of the repo. The second was a copy that got
distributed with the package. This is error prone and unnecessary since
we can package directly from a single file. Since it makes sense for
this file to be exposed at the top level in src/deepforest this puts
the canonical file in that location the removes other files.

Also requires updating some tests to make sure they are getting the config
file from the package itself not the working directory.

Closes #531