Cleanup docs on reading shape files to avoid user confusion (#783)
* Have docs on reading shape files show a standard path

The docs showed reading a file from a tmpdir using the previous
generation of format strings. This is confusing to users (see #781).

Fixes #781

* Remove local shell command; just show csv contents

OS specific shell commands with local path information can be confusing
to users. This removes that command the just shows the output (the
structure of the csv file) which is all the user needs to understand
the behavior of the documented functions.