Allow old configs to be used in new 2.0 checkpoints (#1140)
This commit allows old configs to be used in new 2.0 checkpoints for both deepforest.main and model.CropModel

This uses the num_classes from the checkpoint in the CropModel to avoid the annoying error saying the model isn't created. Added info to docs on compatibility. I also add plateau loss key from new hydra config

If keys from the new config don't exist in the old config, the defaults are used.